Move jobs dir
