+++ api-controllers/ServiceFabrikAdminController.js
@@ -13,7 +13,7 @@ const FabrikBaseController = require('./FabrikBaseController');
 const utils = require('../broker/lib/utils');
 const fabrik = require('../broker/lib/fabrik');
 const bosh = require('../data-access-layer/bosh');
-const ScheduleManager = require('../broker/lib/jobs');
+const ScheduleManager = require('../jobs');
 const BackupReportManager = require('../reports');
 const CONST = require('../common/constants');
 const maintenanceManager = require('../maintenance').maintenanceManager;
