+++ api-controllers/ServiceFabrikApiController.js
@@ -25,7 +25,7 @@ const JsonWebTokenError = jwt.JsonWebTokenError;
 const ContinueWithNext = errors.ContinueWithNext;
 const InternalServerError = errors.InternalServerError;
 const DeploymentAlreadyLocked = errors.DeploymentAlreadyLocked;
-const ScheduleManager = require('../broker/lib/jobs');
+const ScheduleManager = require('../jobs');
 const config = require('../common/config');
 const CONST = require('../common/constants');
 const catalog = require('../common/models').catalog;
