+++ broker/JobScheduler.js
@@ -74,7 +74,7 @@ class JobScheduler {
     this.workerType = `Worker - ${cluster.worker.id} - ${process.pid}`;
     process.on('message', (msg) => this.handleMessage(msg));
     logger.info(`Starting Service Fabrik Batch Job worker: ${cluster.worker.id} - ${process.pid}  @${new Date()}`);
-    require('./lib/jobs');
+    require('../jobs');
     logger.info('Initializing event listeners..');
     utils.initializeEventListener(config.external, 'external');
   }
