+++ jobs/OperationStatusPollerJob.js
@@ -1,15 +1,15 @@
 'use strict';
 
 const _ = require('lodash');
-const logger = require('../../../common/logger');
-const config = require('../../../common/config');
+const logger = require('../common/logger');
+const config = require('../common/config');
 const BaseJob = require('./BaseJob');
 const moment = require('moment');
-const CONST = require('../../../common/constants');
-const errors = require('../../../common/errors');
+const CONST = require('../common/constants');
+const errors = require('../common/errors');
 const ScheduleManager = require('./ScheduleManager');
-const backupStore = require('../../../data-access-layer/iaas').backupStoreForOob;
-const utils = require('../utils');
+const backupStore = require('../data-access-layer/iaas').backupStoreForOob;
+const utils = require('../common/utils');
 const Promise = require('bluebird');
 
 class OperationStatusPollerJob extends BaseJob {
