+++ jobs/ScheduledOobDeploymentBackupJob.js
@@ -2,15 +2,15 @@
 
 const _ = require('lodash');
 const Promise = require('bluebird');
-const logger = require('../../../common/logger');
-const config = require('../../../common/config');
+const logger = require('../common/logger');
+const config = require('../common/config');
 const moment = require('moment');
 const BaseJob = require('./BaseJob');
-const CONST = require('../../../common/constants');
-const errors = require('../../../common/errors');
+const CONST = require('../common/constants');
+const errors = require('../common/errors');
 const NotFound = errors.NotFound;
-const backupStore = require('../../../data-access-layer/iaas').backupStoreForOob;
-const bosh = require('../../../data-access-layer/bosh');
+const backupStore = require('../data-access-layer/iaas').backupStoreForOob;
+const bosh = require('../data-access-layer/bosh');
 const ScheduleManager = require('./ScheduleManager');
 
 class ScheduledOobDeploymentBackupJob extends BaseJob {
