+++ jobs/Scheduler.js
@@ -11,16 +11,16 @@ Promise.promisifyAll([
 const Agenda = require('agenda');
 const CronTime = require('cron').CronTime;
 const humanInterval = require('human-interval');
-const logger = require('../../../common/logger');
-const config = require('../../../common/config');
-const CONST = require('../../../common/constants');
+const logger = require('../common/logger');
+const config = require('../common/config');
+const CONST = require('../common/constants');
 const JobFabrik = require('./JobFabrik');
 const BaseJob = require('./BaseJob');
-const errors = require('../../../common/errors');
+const errors = require('../common/errors');
+const maintenanceManager = require('../maintenance').maintenanceManager;
 const BadRequest = errors.BadRequest;
 const ServiceUnavailable = errors.ServiceUnavailable;
 const ServiceInMaintenance = errors.ServiceInMaintenance;
-const maintenanceManager = require('../../../maintenance').maintenanceManager;
 const MONGO_TO_BE_INITIALIZED = 0;
 const MONGO_INIT_FAILED = 1;
 const MONGO_INIT_SUCCEEDED = 2;
