+++ managers/backup-manager/BackupService.js
@@ -10,7 +10,7 @@ const backupStore = require('../../data-access-layer/iaas').backupStore;
 const utils = require('../../broker/lib/utils');
 const eventmesh = require('../../eventmesh');
 const Agent = require('../../broker/lib/fabrik/Agent');
-const ScheduleManager = require('../../broker/lib/jobs');
+const ScheduleManager = require('../../jobs');
 const CONST = require('../../common/constants');
 const BaseDirectorService = require('../BaseDirectorService');
 const Forbidden = errors.Forbidden;
