+++ test/test_broker/acceptance/service-broker-api-2.0.instances.director.spec.js
@@ -9,7 +9,7 @@ const utils = lib.utils;
 const config = require('../../../common/config');
 const catalog = require('../../../common/models').catalog;
 const fabrik = lib.fabrik;
-const ScheduleManager = require('../../../broker/lib/jobs');
+const ScheduleManager = require('../../../jobs');
 const CONST = require('../../../common/constants');
 const DirectorManager = lib.fabrik.DirectorManager;
 const cloudController = require('../../../data-access-layer/cf').cloudController;
