+++ test/test_broker/acceptance/service-fabrik-api.instances.director.spec.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const Promise = require('bluebird');
 const moment = require('moment');
 const lib = require('../../../broker/lib');
-const ScheduleManager = require('../../../broker/lib/jobs');
+const ScheduleManager = require('../../../jobs');
 const CONST = require('../../../common/constants');
 const apps = require('../support/apps');
 const catalog = require('../../../common/models').catalog;
