+++ test/test_broker/jobs.BackupReaperJobs.spec.js
@@ -4,9 +4,9 @@ const _ = require('lodash');
 const CONST = require('../../common/constants');
 const config = require('../../common/config');
 const moment = require('moment');
-const JobFabrik = require('../../broker/lib/jobs/JobFabrik');
-const BaseJob = require('../../broker/lib/jobs/BaseJob');
-const ScheduleManager = require('../../broker/lib/jobs/ScheduleManager');
+const JobFabrik = require('../../jobs/JobFabrik');
+const BaseJob = require('../../jobs/BaseJob');
+const ScheduleManager = require('../../jobs/ScheduleManager');
 const errors = require('../../common/errors');
 const NotFound = errors.NotFound;
 const backupStore = require('../../data-access-layer/iaas').backupStore;
