+++ test/test_broker/jobs.BaseJob.spec.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const _ = require('lodash');
-const BaseJob = require('../../broker/lib/jobs/BaseJob');
+const BaseJob = require('../../jobs/BaseJob');
 const Repository = require('../../common/db').Repository;
 const CONST = require('../../common/constants');
