+++ test/test_broker/jobs.ScheduleBackupJob.spec.js
@@ -6,9 +6,9 @@ const config = require('../../common/config');
 const moment = require('moment');
 const Promise = require('bluebird');
 const errors = require('../../common/errors');
-const ScheduleManager = require('../../broker/lib/jobs/ScheduleManager');
-const JobFabrik = require('../../broker/lib/jobs/JobFabrik');
-const BaseJob = require('../../broker/lib/jobs/BaseJob');
+const ScheduleManager = require('../../jobs/ScheduleManager');
+const JobFabrik = require('../../jobs/JobFabrik');
+const BaseJob = require('../../jobs/BaseJob');
 const backupStore = require('../../data-access-layer/iaas').backupStore;
 const filename = require('../../data-access-layer/iaas').backupStore.filename;
 const iaas = require('../../data-access-layer/iaas');
